Grand refactoring, fixed every single comment

+++ src/lesson1/task1/Simple.kt
@@ -72,7 +72,7 @@ fun lengthInMeters(sagenes: Int, arshins: Int, vershoks: Int): Double =
  * Вывести значение того же угла в радианах (например, 0.63256).
  */
 fun angleInRadian(grad: Int, min: Int, sec: Int): Double =
-        toRadians(grad.toDouble() + min / 60.0 + sec / 3600.0)
+        toRadians(grad + min / 60.0 + sec / 3600.0)
 
 /**
  * Тривиальная
